rest_backend/requirements/production.txt
Django==2.0
django-material==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
pytz==2017.3
sqlparse==0.2.4
